Optimize!
Instead of keeping an array that's the full board, we only keep the
populated points.  We then know that any point which isn't listed is
blank.

Most Conway patterns have more whitespace than populated cells so this
should optimize both memory usage (because we keep less data in memory)
and CPU usage because we don't have to loop over every cell to determine
which ones might change.